feat(aleo): private list ops part 1
