+++ .changeset/afraid-kings-count.md
@@ -1,0 +1,5 @@
+---
+"@ledgerhq/coin-aleo": minor
+---
+
+aleo private list ops cleanup before extending framework
