+++ libs/coin-modules/coin-aleo/src/logic/index.ts
@@ -5,7 +5,6 @@ export { estimateFees } from "./estimateFees";
 export { getAccountInfo } from "./getAccountInfo";
 export { getBalance } from "./getBalance";
 export { lastBlock } from "./lastBlock";
-export { listOperations } from "./listOperations";
 export { register } from "./register";
 export { validateIntent } from "./validateIntent";
 export { validateAddress } from "./validateAddress";
